functions/requirements.txt
functions-framework==3.*
google-generativeai==0.8.4
openai==1.61.0
scikit-learn==1.6.1
python-dotenv==1.0.1